Various bug fixes to subjects and routing
